requirements.txt
python-telegram-bot==21.6
aiohttp==3.10.5
cachetools==5.5.0
